% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{check_hull_overlap}
\alias{check_hull_overlap}
\title{Check and report convex hull overlap for a fitted GPS object (d = 2 only)}
\usage{
check_hull_overlap(gps)
}
\arguments{
\item{gps}{Object of class \code{"gps_weights"}}
}
\value{
Invisible logical vector of length n (\code{TRUE} = inside hull)
}
\description{
Deprecated in favour of \code{\link[=check_overlap]{check_overlap()}}, which works for any d.
}
