% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_marginal_fit}
\alias{check_marginal_fit}
\title{PIT uniformity plots for marginal distribution fit assessment}
\usage{
check_marginal_fit(gps)
}
\arguments{
\item{gps}{Object of class \code{"gps_weights"}}
}
\value{
A patchwork object (printed and returned invisibly)
}
\description{
Plots probability-integral-transform (PIT) values against Uniform(0,1)
quantiles.  Points on the diagonal indicate a good marginal fit.
}
