% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{.fit_null_marginals}
\alias{.fit_null_marginals}
\title{Compute log marginal density ratios log\link{f_j(T_j)} - log\link{f_j(T_j|X_j)}}
\usage{
.fit_null_marginals(y, margins, marginal_fits, lod_vec, eps)
}
\description{
Compute log marginal density ratios log\link{f_j(T_j)} - log\link{f_j(T_j|X_j)}
}
\keyword{internal}
