% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{.get_copula_family}
\alias{.get_copula_family}
\title{Map a copula type string to VineCopula integer family code(s)}
\usage{
.get_copula_family(copula_type)
}
\description{
Map a copula type string to VineCopula integer family code(s)
}
\keyword{internal}
