% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.impute_censored_copula}
\alias{.impute_censored_copula}
\title{Impute censored treatment values using the fitted copula conditional distribution}
\usage{
.impute_censored_copula(y, cens_mask, lod, mfits, margins, cop, d)
}
\arguments{
\item{y}{n x d treatment matrix (current values; censored cells contain
Beta-initialised or previous-iteration imputed values)}

\item{cens_mask}{n x d logical matrix; TRUE = censored}

\item{lod}{numeric vector of length d}

\item{mfits}{list of marginal fit objects from \code{.fit_marginal()}}

\item{margins}{character vector of length d}

\item{cop}{BiCop object (family, par, par2)}

\item{d}{number of treatments (must be 2)}
}
\value{
Updated n x d matrix with imputed values for censored cells
}
\description{
Draws from P(T_j | T_j < LOD_j, T_{-j}, X) by inverting the copula
h-function sequentially (j = 1 then j = 2).  After imputing T_1, the
updated value is immediately used as the conditioning partner when
imputing T_2 (Algorithm A2).
}
\keyword{internal}
