% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.init_censored}
\alias{.init_censored}
\title{Initialise censored cells with Beta(2,1) draws scaled to [0, LOD)}
\usage{
.init_censored(y, cens_mask, lod)
}
\arguments{
\item{y}{n x d treatment matrix}

\item{cens_mask}{n x d logical censoring mask}

\item{lod}{numeric vector of length d}
}
\value{
y with censored cells replaced by Beta(2,1) * LOD initialisations
}
\description{
Initialise censored cells with Beta(2,1) draws scaled to [0, LOD)
}
\keyword{internal}
