% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_chain.R
\name{.pool_gps_diagnostics}
\alias{.pool_gps_diagnostics}
\title{Average GPS diagnostics across chains (fallback for mode = "gps")}
\usage{
.pool_gps_diagnostics(valid_chains, M)
}
\description{
Average GPS diagnostics across chains (fallback for mode = "gps")
}
\keyword{internal}
