% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.run_imputation_loop}
\alias{.run_imputation_loop}
\title{Run the iterative copula imputation loop}
\usage{
.run_imputation_loop(
  y,
  data,
  margins,
  copula_type,
  cens_mask,
  lod,
  params,
  df,
  n_iter,
  conv_tol,
  d,
  verbose
)
}
\arguments{
\item{y}{n x d treatment matrix (censored cells initialised with Beta draws)}

\item{data}{internal data list (for design matrices)}

\item{margins}{character vector of length d}

\item{copula_type}{copula family string}

\item{cens_mask}{n x d logical censoring mask}

\item{lod}{numeric vector of length d}

\item{params}{optional user-supplied parameters list}

\item{df}{degrees of freedom for t margin}

\item{n_iter}{maximum iterations}

\item{conv_tol}{convergence tolerance}

\item{d}{number of treatments}

\item{verbose}{print iteration messages}
}
\value{
List: y (final imputed matrix), marginal_fits, n_iter_done
}
\description{
Fits marginals and re-imputes censored values until the maximum relative
change in imputed cells drops below \code{conv_tol} or \code{n_iter} is reached.
}
\keyword{internal}
