% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_utils.R
\name{.vine_type_code}
\alias{.vine_type_code}
\title{Map vine type string to VineCopula integer type code}
\usage{
.vine_type_code(vine_type)
}
\description{
Map vine type string to VineCopula integer type code
}
\keyword{internal}
