% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{.weighted_cor_matrix}
\alias{.weighted_cor_matrix}
\title{Vectorised weighted correlation between two numeric matrices}
\usage{
.weighted_cor_matrix(X, Y, w = NULL)
}
\description{
Vectorised weighted correlation between two numeric matrices
}
\keyword{internal}
