% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{get_ate_table}
\alias{get_ate_table}
\title{ATE or OR estimates as a gt table}
\usage{
get_ate_table(ate)
}
\arguments{
\item{ate}{Object of class \code{"gps_ate"}}
}
\value{
A gt table
}
\description{
ATE or OR estimates as a gt table
}
