% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_copula_params}
\alias{get_copula_params}
\title{Extract copula parameters as a list}
\usage{
get_copula_params(gps)
}
\arguments{
\item{gps}{Object of class \code{"gps_weights"}}
}
\value{
List with empirical and parametric copula details
}
\description{
Extract copula parameters as a list
}
