% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_copula_table}
\alias{get_copula_table}
\title{Copula parameters as a gt table}
\usage{
get_copula_table(gps)
}
\arguments{
\item{gps}{Object of class \code{"gps_weights"}}
}
\value{
A gt table
}
\description{
Copula parameters as a gt table
}
