% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_marginal_params}
\alias{get_marginal_params}
\title{Extract marginal distribution parameters from a GPS object}
\usage{
get_marginal_params(gps)
}
\arguments{
\item{gps}{Object of class \code{"gps_weights"}}
}
\value{
Named list of per-treatment parameter summaries
}
\description{
Extract marginal distribution parameters from a GPS object
}
