% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.gps_drf}
\alias{plot.gps_drf}
\title{Plot a GPS dose-response function}
\usage{
\method{plot}{gps_drf}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"gps_drf"} from \code{\link[=drf]{drf()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a GPS dose-response function
}
